+++ docs/changelog.md
@@ -286,6 +286,9 @@ Many feature gates are removed in order to make most features make always availa
 
 #### ascii
 - rename `Ascii` to `AsciiDigits`.
+- update `AsciiDigits`:
+  - new const: `MAX_DIGITS_16`.
+  - new methods: `count_digits16`, `digits16`, `digit_at_power16`.
   - rename const: `MAX_DIGITS` to `MAX_DIGITS_10`.
   - rename methods:
     - `calc_digit` to `digit_at_power10`.